Square 1
This is the initial commit, as such, it mostly works, there
are things that are good and things that are not so good.
With time, the good things will hopefully get more numerous
while the not so good things lessen, assuming they actually
cause enough issues that they need to be removed via more good things.